removing underscore
