+++ pkg/identity/ciprovider/principal.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ci_provider
+package ciprovider
 
 import (
 	"context"
